fix(api): fix controller response error

+++ apps/api/src/room/controllers/admin.controller.ts
@@ -5,6 +5,7 @@ import {
   Get,
   HttpException,
   HttpStatus,
+  NotFoundException,
   Param,
   ParseIntPipe,
   Post,
@@ -12,6 +13,7 @@ import {
   Query,
   Req,
   Res,
+  UploadedFiles,
   UseGuards
 } from '@nestjs/common'
 import { IResponseData, IResponsePaginate, RoomCreateDto, RoomUpdateDto } from '@riverrun/interface'
@@ -26,7 +28,12 @@ export class RoomAdminController {
   constructor(private readonly roomService: RoomService) {}
 
   @Post()
-  async create(@Req() req: Request, @Res() res: Response, @Body() body: RoomCreateDto) {
+  async create(
+    @Req() req: Request,
+    @Res() res: Response,
+    @UploadedFiles() files: Array<Express.Multer.File>,
+    @Body() body: RoomCreateDto
+  ) {
     try {
       const query = await this.roomService.create(body)
       const response: IResponseData<string> = {
@@ -50,43 +57,31 @@ export class RoomAdminController {
     @Query('page') page: number,
     @Query('limit') limit: number
   ) {
-    try {
-      const currentPage = page || 1
-      const perPage = limit || 10
-      const query = await this.roomService.findAll(currentPage, perPage)
-      const total = await this.roomService.count()
-      const response: IResponsePaginate<Room[]> = {
-        success: true,
-        total: total,
-        currentPage: currentPage,
-        perPage: perPage,
-        data: query
-      }
-      res.status(HttpStatus.OK).json(response)
-    } catch (error) {
-      const message = {
-        message: error.message
-      }
-      throw new HttpException(message, HttpStatus.BAD_REQUEST)
+    const currentPage = page || 1
+    const perPage = limit || 10
+    const query = await this.roomService.findAll(currentPage, perPage)
+    const total = await this.roomService.count()
+    const response: IResponsePaginate<Room[]> = {
+      success: true,
+      total: total,
+      currentPage: currentPage,
+      perPage: perPage,
+      data: query
     }
+    res.status(HttpStatus.OK).json(response)
   }
 
   @Get(':id')
   async findByID(@Req() req: Request, @Res() res: Response, @Param('id', ParseIntPipe) id: number) {
-    try {
-      const query = await this.roomService.findByID(id)
-      const response: IResponseData<Room> = {
-        data: query,
-        success: true
-      }
-      res.status(HttpStatus.OK).json(response)
-    } catch (error) {
-      const msg: IResponseData<string> = {
-        message: error?.message,
-        success: false
-      }
-      throw new HttpException(msg, HttpStatus.FORBIDDEN)
+    const query = await this.roomService.findByID(id)
+    if (!query) {
+      throw new NotFoundException('id not found')
     }
+    const response: IResponseData<Room> = {
+      data: query,
+      success: true
+    }
+    res.status(HttpStatus.OK).json(response)
   }
 
   @Put(':id')
@@ -94,40 +89,33 @@ export class RoomAdminController {
     @Req() req: Request,
     @Res() res: Response,
     @Param('id', ParseIntPipe) id: number,
+    @UploadedFiles() files: Array<Express.Multer.File>,
     @Body() body: RoomUpdateDto
   ) {
-    try {
-      await this.roomService.update(id, body)
-      const query = await this.roomService.findByID(id)
-      const response: IResponseData<Room> = {
-        data: query,
-        success: true
-      }
-      res.status(HttpStatus.OK).json(response)
-    } catch (error) {
-      const msg: IResponseData<string> = {
-        message: error?.message,
-        success: false
-      }
-      throw new HttpException(msg, HttpStatus.FORBIDDEN)
+    const query = await this.roomService.findByID(id)
+    if (!query) {
+      throw new NotFoundException('id not found')
     }
+    await this.roomService.update(id, body)
+    await this.roomService.findByID(id)
+    const response: IResponseData<Room> = {
+      data: query,
+      success: true
+    }
+    res.status(HttpStatus.OK).json(response)
   }
 
   @Delete(':id')
   async remove(@Req() req: Request, @Res() res: Response, @Param('id', ParseIntPipe) id: number) {
-    try {
-      await this.roomService.remove(id)
-      const response: IResponseData<string> = {
-        message: 'Delete successfully',
-        success: true
-      }
-      res.status(HttpStatus.OK).json(response)
-    } catch (error) {
-      const msg: IResponseData<string> = {
-        message: error?.message,
-        success: false
-      }
-      throw new HttpException(msg, HttpStatus.FORBIDDEN)
+    const query = await this.roomService.findByID(id)
+    if (!query) {
+      throw new NotFoundException('id not found')
     }
+    await this.roomService.remove(id)
+    const response: IResponseData<string> = {
+      message: 'Delete successfully',
+      success: true
+    }
+    res.status(HttpStatus.OK).json(response)
   }
 }
